% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sktcal.R
\name{sktcal}
\alias{sktcal}
\title{skin temp calibration between biolab and anslab}
\usage{
sktcal(biolab1, anslab1, biolab2, anslab2)
}
\arguments{
\item{biolab1}{first value in from biolab}

\item{anslab1}{corresponding anslab value to first value from biolab}

\item{biolab2}{second value in from biolab}

\item{anslab2}{corresponding anslab value to second value from biolab}
}
\description{
exclude based on user input sd from the mean
}
